backend/pyproject.toml
[tool.poetry]
name = "centralized-advertising-ai-analytics"
version = "0.1.0"
description = ""
authors = ["dotpep <pepelnine@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.115.6"}
sqlalchemy = "^2.0.36"
sqlalchemy-utils = "^0.41.2"
psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
alembic = "^1.12.0"
faker = "^20.0.0"
pandas = "^2.1.0"
numpy = "^1.26.0"
seaborn = "^0.13.0"
matplotlib = "^3.8.0"
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
